pyproject.toml
[project]
name = "patientjournals"
version = "v1.0"
description = "The project creates a structure for transcribing Patient Journals. Project is developed by Lucas Mortensen @ PandemiX."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.57.1",
    "google-genai>=1.57.0",
    "google-cloud-storage>=2.19.0",
    "invoke>=2.2.0",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pypdfium2>=4.30.0",
    "pydantic>=2.12.5",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
